[Encode] Fix a hang in the free kernel driver
Ensure the BufferMutex in the DdiMedia_MapBufferInternal function gets
unlocked even if there is a failure in a critical section.
MMC support is disabled in the free kernel open-linux-driver build. A
check for MMC support inside a critical section was implemented using
the DDI_CHK_RET macro that triggers exiting from the function upon the
check failing. Bailing out without unlocking the mutex led to a hang.

Change-Id: I1488b323df7b8b9b8bed54109e5837d70345d06e