[Encode][Issuefix] Fix a ULT assert issue
[Encode][Issuefix] Fix a ULT assert issue
When in the following condition:
1. Get command buffer;
2. Switch to another context;
3. Another context gets, submits, returns command buffer;
4. Switch back to current conetxt;
5. Current context submits, returns command buffer.
The last return will trigger an assetion because previous command buffer status was lost caused by context switch.
This change aims to support this usage by adding a flag to indicate whether it is context switch back, if so command buffer will not be reset.

Change-Id: I87da49e2e481846a1b1a0f3e81f0aa1b456ee275